requirements.txt
openai>=1.0.0
groq>=0.4.0
google-generativeai>=0.3.0
cohere>=5.0.0
huggingface-hub>=0.19.0
requests>=2.31.0
python-dotenv>=1.0.0
ollama>=0.1.0
